backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
openai
langfuse
pandas
openpyxl
requests

